+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   declaring methods and properties
 - User's ID for signal messages (this was added as an option in SHV standard)
 - `SHVBase._ls_node_for_path` helper
+- Support for alerts in `SHVDevice`
+- `RpcAlert` as abstraction for device alerts
 
 ### Changed
 - "Simple" API was renamed to SHV and thus `SimpleClient` is now `SHVClient`